requirements.txt
streamlit==1.46.0
python-kucoin==2.2.0
pandas==2.3.0
numpy==2.3.1
plotly==6.1.2
python-dotenv==1.1.0
ccxt==4.4.90
requests==2.32.4
websocket-client==1.8.0
schedule==1.2.2
pytest==8.4.1